+++ main.go
@@ -3,21 +3,18 @@ package main
 import (
 	"net/http"
 
+	app "github.com/ainsleybc/neighbourly/app"
 	r "github.com/dancannon/gorethink"
 )
 
-var session *r.Session
-
 func main() {
 
-	session, _ = r.Connect(r.ConnectOpts{
+	session, _ := r.Connect(r.ConnectOpts{
 		Address:  "localhost:28015",
 		Database: "neighbourly",
 	})
 
-	router := NewRouter()
-
-	router.RegisterHandler("feed add", addFeed)
+	router := app.NewRouter(session)
 
 	http.Handle("/", router)
 	http.ListenAndServe(":4000", nil)
